# packages/schemas/openapi.v1.yaml
openapi: 3.1.0
info:
  title: MARID Control Plane API
  version: 0.3.0
  description: Tenant-scoped control plane. Execution records do not execute tools.
servers:
  - url: http://localhost:5056
paths:
...
  /v1/evidence:
    post:
      summary: Stream one raw evidence object into private storage
      description: Query metadata includes source, sourceEventId, sourceAt (UTC), sensorIdentity, parserVersion, and optional retentionClass. Maximum body is 64 MiB. A retry with the same source ID and bytes is idempotent.
      security: [{ bearerAuth: [] }]
      x-required-scope: evidence.ingest
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '201': { description: Evidence stored with SHA-256 digest }
        '200': { description: Identical evidence already exists }
        '400': { description: Invalid metadata or empty body }
        '403': { description: Missing tenant ingest role or scope }
        '409': { description: Source ID belongs to different evidence }
        '413': { description: Body is too large }
        '503': { description: Object storage is unavailable }
  /v1/evidence/{id}:
    get:
      summary: Read tenant evidence metadata without exposing its object key
      security: [{ bearerAuth: [] }]
      x-required-scope: evidence.read
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Tenant evidence metadata }
        '403': { description: Missing tenant role or scope }
        '404': { description: No evidence with this ID in the token tenant }
  /v1/evidence/{id}/content:
    get:
      summary: Verify digest and stream tenant raw evidence
      security: [{ bearerAuth: [] }]
      x-required-scope: evidence.download
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Verified evidence bytes }
        '403': { description: Missing tenant role or scope }
        '404': { description: No evidence with this ID in the token tenant }
        '503': { description: Storage unavailable or integrity check failed }
  /v1/exports:
    post:
      summary: Queue an asynchronous tenant export for an explicit UTC interval
      security: [{ bearerAuth: [] }]
      x-required-scope: exports.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [periodStart, periodEnd, idempotencyKey]
              properties:
                periodStart: { type: string, format: date-time }
                periodEnd: { type: string, format: date-time }
                idempotencyKey: { type: string, minLength: 8, maxLength: 120 }
      responses:
        '202': { description: Job queued }
        '200': { description: Same idempotent job already exists }
        '400': { description: Invalid UTC interval }
        '403': { description: Missing tenant role or scope }
        '409': { description: Idempotency key belongs to a different interval }
        '503': { description: Worker or object storage is unavailable }
    get:
      summary: List the latest 100 export jobs in the token tenant
      security: [{ bearerAuth: [] }]
      x-required-scope: exports.read
      responses:
        '200': { description: Tenant export jobs }
  /v1/exports/schedule:
    get:
      summary: Read the tenant monthly UTC export schedule
      security: [{ bearerAuth: [] }]
      x-required-scope: exports.read
      responses:
        '200': { description: Schedule }
        '404': { description: No schedule configured }
    put:
      summary: Set day 1 to 28 for monthly export generation
      security: [{ bearerAuth: [] }]
      x-required-scope: exports.admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled, dayOfMonth]
              properties:
                enabled: { type: boolean }
                dayOfMonth: { type: integer, minimum: 1, maximum: 28 }
      responses:
        '204': { description: Schedule updated and audited }
        '403': { description: Requires a human incident manager }
  /v1/exports/{id}:
    get:
      summary: Read one tenant export job
      security: [{ bearerAuth: [] }]
      x-required-scope: exports.read
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Job status }
        '404': { description: No job with this ID in the token tenant }
  /v1/exports/{id}/artifacts:
    get:
      summary: List artifact IDs, names, sizes, and hashes for a tenant job
      security: [{ bearerAuth: [] }]
      x-required-scope: exports.read
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Artifact metadata }
        '404': { description: No job with this ID in the token tenant }
  /v1/exports/{id}/manifest:
    get:
      summary: Verify and stream the completed export manifest
      security: [{ bearerAuth: [] }]
      x-required-scope: exports.download
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: SHA-256 verified manifest JSON }
        '404': { description: No completed job with this ID in the token tenant }
        '503': { description: Storage or integrity failure }
  /v1/exports/{id}/artifacts/{artifactId}:
    get:
      summary: Verify and stream one completed tenant export artifact
      security: [{ bearerAuth: [] }]
      x-required-scope: exports.download
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: artifactId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: SHA-256 verified JSON, Markdown, or NDJSON artifact }
        '404': { description: No artifact with this ID in this tenant job }
        '503': { description: Storage or integrity failure }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT